Refactor into existing Workspace component
