+++ public/pages/workflow_detail/workspace/reactflow-styles.scss
@@ -8,3 +8,9 @@
   flex-grow: 1;
   height: 100%;
 }
+
+.workspace {
+  width: 50vh;
+  height: 50vh;
+  padding: 0;
+}
